requirements.txt
typer[all]==0.9.0
aiohttp==3.8.4
requests==2.31.0
python-dotenv==1.0.0
PyYAML==6.0
Pillow==10.0.1
pandas==2.2.2
openai==1.0.0
APScheduler==3.10.1
tenacity==8.2.2
beautifulsoup4>=4.12.0
selenium
webdriver-manager